shooter: more cleanup and update extract scripts

+++ unzip-files.sh
@@ -1,4 +1,11 @@
-#!/bin/sh
+#/bin/sh
+
+UPDATE_PACKAGE=shooter-update.zip
+
+if [ "x$UPDATE_PACKAGE" = "x" ]; then
+    echo You must specify the update.zip as first argument
+    exit
+fi
 
 BASE=../../../vendor/htc/shooter/proprietary
 rm -rf $BASE/*
@@ -8,7 +15,7 @@ for FILE in `egrep -v '(^#|^$)' proprietary-files.txt`; do
   if [ ! -d $BASE/$DIR ]; then
     mkdir -p $BASE/$DIR
   fi
-  adb pull $FILE $BASE/$FILE
+  unzip -j -o ${UPDATE_PACKAGE} $FILE -d $BASE/$DIR
 done
 
 ./setup-makefiles.sh
